requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-experimental==0.3.4
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
openai==1.79.0
chromadb==1.0.10
httpx==0.28.1
httpx-sse==0.4.
typing_extensions==4.13.2
pdfminer.six==20250506
unstructured==0.17.2
pi-heif>=0.8.0
unstructured[local-inference]>=0.12.0